feat: improve Swap Button
