+++ packages/kit/src/views/Swap/index.tsx
@@ -8,7 +8,6 @@ import { useActiveWalletAccount } from '../../hooks/redux';
 import { reset } from '../../store/reducers/swap';
 
 import { useSwapActionHandlers, useSwapEnabled } from './hooks/useSwap';
-import NoWallet from './NoWallet';
 import { refs } from './refs';
 import SwapContent from './SwapContent';
 import SwapHeader from './SwapHeader';
@@ -48,7 +47,6 @@ const Swap = () => {
     <ScrollView>
       <SwapHeader />
       <SwapTransactions />
-      <NoWallet />
       <SwapContent />
       <SwapItems />
       <SwapUpdator />
